requirements.txt
cachetools
fastapi
pydantic
slowapi
uvicorn
fastapi-cache2[memcache]
sqlalchemy[asyncio]
aiosqlite
python-dotenv